requirements.txt
fastapi
uvicorn[standard]
httpx
Faker 
requests
stripe
python-dotenv
